+++ releasenotes.md
@@ -1,5 +1,9 @@
 ## vNext
 
+## 0.20.2
+
+- Fixed a potential memory Leak for missing dispose in CancellationTokenSource and ReaderWriterLockSlim.
+
 ## 0.20.1
 
 - Upgraded MongoDb Driver to 3.4.2.